nlp_analysis/requirements.txt
spacy>=3.5,<3.8
nltk>=3.8
scikit-learn>=1.2
gensim>=4.3
matplotlib>=3.7
seaborn>=0.12
wordcloud>=1.9
pyyaml>=6.0
numpy>=1.24
scipy>=1.10
python-docx>=0.8.11
